// src/types/real-type-types.type.ts
export type RealTypeTypes =
  | 'Undefined'
  | 'Null'
  | 'RegExp'
  | 'String'
  | 'BigInt'
  | 'Date'
  | 'Array'
  | 'Symbol'
  | 'Int8Array'
  | 'Uint8Array'
  | 'Uint8ClampedArray'
  | 'Int16Array'
  | 'Uint16Array'
  | 'Int32Array'
  | 'Uint32Array'
  | 'Float32Array'
  | 'Float64Array'
  | 'BigInt64Array'
  | 'BigUint64Array'
  | 'Map'
  | 'WeakMap'
  | 'Set'
  | 'WeakSet'
  | 'ArrayBuffer'
  | 'SharedArrayBuffer'
  | 'Function'
  | 'Generator'
  | 'GeneratorFunction'
  | 'Arguments'
  | 'Number'
  | 'Object'
  | 'Error'
  | 'TypeError'
  | 'EvalError'
  | 'SyntaxError'
  | 'URIError'
  | 'RangeError'
  | 'ReferenceError'
  | 'NaN'
  | 'Infinity'
  | 'SafeInteger'
  | 'UnpreciseInteger'
  | 'Float';